+++ pkg/idp/provider/azure/azure_test.go
@@ -4,6 +4,7 @@ import (
 	"giantswarm/dex-operator/pkg/idp/provider"
 	"strconv"
 	"testing"
+	"time"
 
 	"github.com/go-logr/logr"
 	"github.com/microsoftgraph/msgraph-sdk-go/models"
@@ -128,6 +129,45 @@ func TestNewConfig(t *testing.T) {
 			}
 			if err == nil && tc.expectError {
 				t.Fatalf("Expected an error, got success.")
+			}
+		})
+	}
+}
+
+func TestSecretExpired(t *testing.T) {
+	testCases := []struct {
+		name           string
+		expirationDate time.Time
+		expired        bool
+	}{
+		{
+			name:           "case 0",
+			expirationDate: time.Now(),
+			expired:        true,
+		},
+		{
+			name:           "case 1",
+			expirationDate: time.Now().Add(7 * 24 * time.Hour),
+			expired:        true,
+		},
+		{
+			name:           "case 2",
+			expirationDate: time.Now().Add(14 * 24 * time.Hour),
+			expired:        false,
+		},
+		{
+			name:           "case 3",
+			expirationDate: time.Now().Add(-1 * 24 * time.Hour),
+			expired:        true,
+		},
+	}
+
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			s := models.NewPasswordCredential()
+			s.SetEndDateTime(&tc.expirationDate)
+			if secretExpired(s) != tc.expired {
+				t.Fatalf("Expected %v, got %v", tc.expired, secretExpired(s))
 			}
 		})
 	}