+++ pkg/idp/provider/azure/azure.go
@@ -37,6 +37,12 @@ type Config struct {
 	ClientSecret string
 }
 
+type config struct {
+	tenantID     string
+	clientID     string
+	clientSecret string
+}
+
 func New(p provider.ProviderCredential, log *logr.Logger) (*Azure, error) {
 
 	// get configuration from credentials
@@ -47,6 +53,7 @@ func New(p provider.ProviderCredential, log *logr.Logger) (*Azure, error) {
 
 	var client *msgraphsdk.GraphServiceClient
 	{
+
 		cred, err := azidentity.NewClientSecretCredential(c.TenantID, c.ClientID, c.ClientSecret, nil)
 		if err != nil {
 			return nil, microerror.Mask(err)
@@ -65,6 +72,7 @@ func New(p provider.ProviderCredential, log *logr.Logger) (*Azure, error) {
 		}
 	}
 	return &Azure{
+
 		Name:         key.GetProviderName(p.Owner, p.Name),
 		Log:          log,
 		Type:         ProviderConnectorType,