+++ CHANGELOG.md
@@ -7,7 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
 - Add use of runtime/default seccomp profile.
+
+### Changed
+
+- Added a new config parameter for issuer address
+- Adjusted creation of callback URL to prefer issuer address over base domain if possibe
 
 ## [0.1.4] - 2022-12-21