+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add github provider
 - Add use of runtime/default seccomp profile.
 
+### Changed
+
+- Added a new config parameter for issuer address
+- Adjusted creation of callback URL to prefer issuer address over base domain if possible
+
 ## [0.1.4] - 2022-12-21
 
 ### Added