+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `setup` package with functionality to create dex-operator credentials
 - Add credential creation, update and cleanup functionality to azure provider
 - Add github provider
+
+## [0.2.0] - 2023-02-27
+
+### Added
+
 - Add use of runtime/default seccomp profile.
 
 ### Changed
@@ -65,7 +70,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added azure active directory provider.
 - Added initial implementation of the dex operator.
 
-[Unreleased]: https://github.com/giantswarm/dex-operator/compare/v0.1.4...HEAD
+[Unreleased]: https://github.com/giantswarm/dex-operator/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/giantswarm/dex-operator/compare/v0.1.4...v0.2.0
 [0.1.4]: https://github.com/giantswarm/dex-operator/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/giantswarm/dex-operator/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/giantswarm/dex-operator/compare/v0.1.1...v0.1.2